server/lambda/services/shared/requirements.txt
# Shared dependencies for all Lambda functions

# Database
psycopg2-binary>=2.9.9

# JWT Authentication
PyJWT>=2.8.0

# AWS SDK (included in Lambda runtime, but needed for local dev)
boto3>=1.34.0

# Typing extensions for Python 3.11
typing-extensions>=4.9.0
